requirements.txt
django==4.2.4
requests==2.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
